Split the admin.py file.
